Experimenting with different code algorithms. NONE ARE WORKING. THIS COMMIT IS JUNK
The issue is that the outer virtual boundary Face Labels and the underlying FeatureIds
do not agree. I think we may need to store the actual index offset into the
FeatureIds array and then also store a reference to the FeatureIds array so that
we can get the data consistently.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>